requirements.txt
appnope==0.1.0
backcall==0.1.0
blessed==1.14.2
botocore==1.10.4
branca==0.2.0
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.7
cycler==0.10.0
decorator==4.2.1
descartes==1.1.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
falcon==1.4.1
Fiona==1.7.11.post1
folium==0.5.0
geopandas==0.3.0
gunicorn==19.7.1
hug==2.4.0
idna==2.6
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
munch==2.3.1
numpy==1.14.2
pandas==0.22.0
parso==0.1.1
pathspec==0.5.5
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
pyproj==1.9.5.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2018.4
PyYAML==3.12
requests==2.18.4
searchtweets==1.7.0
semantic-version==2.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
tweet-parser==1.13.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.47.0
